+++ Package.swift
@@ -14,7 +14,7 @@ let package = Package(
         .target(name: "PatternSpaceSDKClient", dependencies: ["PatternSpaceSDKCore"]),
         .target(name: "PatternSpaceSDKServer", dependencies: ["PatternSpaceSDKCore"]),
         .testTarget(name: "PatternSpaceSDKCoreTests",   dependencies: ["PatternSpaceSDKCore"],   path: "Tests/PatternSpaceSDKCoreTests"),
-        .testTarget(name: "PatternSpaceSDKServerTests", dependencies: ["PatternSpaceSDKServer"], path: "Tests/PatternSpaceSDKServerTests"),
+        .testTarget(name: "PatternSpaceSDKServerTests", dependencies: ["PatternSpaceSDKServer", "PatternSpaceSDKClient"], path: "Tests/PatternSpaceSDKServerTests"),
         .testTarget(name: "PatternSpaceSDKClientTests", dependencies: ["PatternSpaceSDKClient"], path: "Tests/PatternSpaceSDKClientTests"),
     ]
 )